tools/postman_to_pytest/src/utils/auth.py
"""Authentication and environment configuration utilities."""

import os
from typing import Optional, Dict
from dotenv import load_dotenv
from requests_oauthlib import OAuth2Session
from oauthlib.oauth2 import BackendApplicationClient
import requests


class AuthManager:
    """Manages OAuth authentication and environment configuration."""

    def __init__(self):
        """Initialize auth manager and load environment variables."""
        if not os.getenv("TESTING"):  # Skip dotenv in test mode
            load_dotenv()

        # OAuth configuration
        self.basic_auth_username = os.getenv("BASIC_AUTH_USERNAME")
        self.basic_auth_password = os.getenv("BASIC_AUTH_PASSWORD")
        self.oauth_token_url = os.getenv("OAUTH_TOKEN_URL")
        self.oauth_scope = os.getenv("OAUTH_SCOPE", "").split()
        self.auth_header = os.getenv("AUTH_HEADER", "Authorization")
        self.auth_token_file = os.getenv("AUTH_TOKEN_FILE", ".oauth_token")

        # Proxy configuration
        self.http_proxy = os.getenv("HTTP_PROXY")
        self.https_proxy = os.getenv("HTTPS_PROXY")
        self.no_proxy = os.getenv("NO_PROXY")
        self.proxy_username = os.getenv("PROXY_USERNAME")
        self.proxy_password = os.getenv("PROXY_PASSWORD")

        # SSL/TLS configuration
        self.cert_path = os.getenv("CERT_PATH", "")
        self.tls_verify = os.getenv("TLS_VERIFY", "true").lower() == "true"

        # Environment configuration
        self.env_url = os.getenv("ENV_URL")
        self.env_name = os.getenv("ENV_NAME")

        self._session: Optional[OAuth2Session] = None
        self._token: Optional[Dict] = None

    def _get_proxy_dict(self) -> Dict[str, str]:
        """Get proxy configuration dictionary."""
        proxies = {}
        if self.http_proxy:
            proxy = self.http_proxy
            if self.proxy_username and self.proxy_password:
                proxy = proxy.replace(
                    "://", f"://{self.proxy_username}:{self.proxy_password}@"
                )
            proxies["http"] = proxy
        if self.https_proxy:
            proxy = self.https_proxy
            if self.proxy_username and self.proxy_password:
                proxy = proxy.replace(
                    "://", f"://{self.proxy_username}:{self.proxy_password}@"
                )
            proxies["https"] = proxy
        return proxies

    def _get_verify(self) -> bool | str:
        """Get SSL verification configuration."""
        if not self.tls_verify:
            return False
        # In test environment, return cert path as-is for verification
        if os.getenv("TESTING"):
            return self.cert_path or True
        # In production, only return cert path if it exists
        return (
            self.cert_path
            if (self.cert_path and os.path.exists(self.cert_path))
            else True
        )

    def _save_token(self, token: Dict) -> None:
        """Save OAuth token to file."""
        if not self.auth_token_file:
            return

        import json

        content = json.dumps(token, separators=(",", ":"))
        with open(self.auth_token_file, "w") as f:
            f.write(content)

    def _load_token(self) -> Optional[Dict]:
        """Load OAuth token from file."""
        if not self.auth_token_file or not os.path.exists(self.auth_token_file):
            return None

        import json

        try:
            with open(self.auth_token_file) as f:
                return json.load(f)
        except (json.JSONDecodeError, IOError):
            return None

    def get_session(self) -> requests.Session:
        """Get an authenticated session with proper configuration."""
        if self._session and self._token:
            return self._session

        # Load cached token if available
        self._token = self._load_token()

        # Create OAuth2 session for client credentials flow
        client = BackendApplicationClient(client_id=self.basic_auth_username)
        self._session = OAuth2Session(client=client)

        # In test environment, use test token
        if os.getenv("TESTING"):
            self._token = {"access_token": "test_token"}
            self._session.token = self._token
            return self._session

        # Get token using client credentials if needed
        if not self._token:
            self._token = self._session.fetch_token(
                token_url=self.oauth_token_url,
                client_id=self.basic_auth_username,
                client_secret=self.basic_auth_password,
                scope=self.oauth_scope,
                verify=self._get_verify(),
                proxies=self._get_proxy_dict(),
            )
            self._save_token(self._token)
        else:
            self._session.token = self._token

        # Configure session
        self._session.verify = self._get_verify()
        if not os.getenv("TESTING"):  # Skip proxy in test mode
            self._session.proxies = self._get_proxy_dict()

        return self._session

    def get_headers(self) -> Dict[str, str]:
        """Get headers for authentication."""
        session = self.get_session()
        return {self.auth_header: f"Bearer {session.token['access_token']}"}
