tools/postman_to_pytest/requirements.txt
# Core dependencies
pytest>=7.0.0
pytest-dependency>=0.5.1
pyyaml>=6.0.0
click>=8.0.0
requests>=2.28.0
python-dotenv>=1.0.0
requests-oauthlib>=1.3.0

# Development dependencies
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0
